requirements.txt
aiogram==3.13.1
google-api-python-client>=2.111.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
python-dotenv>=1.0.0
